* os_AtomicAccess.h: ifdef PLATFORM_os_AtomicAccess_h, then includes PLATFORM_os_AtomicAccess.h: Special cases platform-depending.   * os_time.h: ifdef PLATFORM_os_time_h, then includes PLATFORM_os_time_h   * ReflMemAccessJc.c: Test nullpointer to prevent crash   * ReflectionJc.c: Using of getAddrElement_FieldJc() instead getMemoryAddress_FieldJc to access getInt_FieldJc etc. for correct supporting arrays.

+++ source/InspcJ2c/ClassContent_Inspc.c
@@ -833,7 +833,7 @@ int32 getSetValueByPath_ClassContent_Inspc(ClassContent_Inspc_s* ythis, struct I
                   
                   
                   setValue = getShort_SetValue_InspcDataExchangeAccess_Inspc(accSetValue, _thCxt);
-                  value = setBitfield_FieldJc(theField, theObject, setValue, "I", idx);
+                  value = (int16)setBitfield_FieldJc(theField, theObject, setValue, "I", idx);
                 }
                 else 
                 {
